Move the distribute orders feature to a separate file
